+++ .vscode/settings.json
@@ -2,6 +2,7 @@
     "cmake.configureOnOpen": true,
     "files.associations": {
         "stdint.h": "c",
-        "rf_pico.h": "c"
+        "rf_pico.h": "c",
+        "cyw43_arch.h": "c"
     }
 }
