+++ inc/rf_pico.h
@@ -4,16 +4,6 @@
 #include "pico/stdlib.h"
 #include "rf_device.h"
 
-// Different LEDS for different boards
-#ifdef USING_PICO_W
-#include "pico/cyw43_arch.h"
-#define LED_PIN CYW43_WL_GPIO_LED_PIN
-#define LED_PIN_PUT cyw43_arch_gpio_put
-#else
-#define LED_PIN 25
-#define LED_PIN_PUT gpio_put
-#endif
-
 #define GPIO_PIN 22
 
 /**
@@ -41,55 +31,13 @@ typedef struct
     repeating_timer_t timer;
 } rf_pico_receiver;
 
-
-//uint64_t pico_get_timestamp_us_callback();
-
 /**
- * Sets the signal for pico_tx.
+ * @brief Initializes the RF Pico transmitter.
  *
- * @param is_high The signal value to set. True for high, false for low.
- * @param user_data A pointer to user-defined data.
+ * @param self The RF Pico transmitter.
+ * @param gpio The GPIO pin to use for the transmitter.
  */
-void pico_tx_set_signal(uint8_t is_high, void* user_data);
-/**
- * @brief Callback function for reading data in the Pico device.
- *
- * This function is called when data needs to be read from the Pico device.
- * It takes a void pointer as an argument, which can be used to pass user-defined data.
- *
- * @param user_data A pointer to user-defined data.
- */
-void pico_data_read_callback(void *user_data);
-/**
- * @brief Callback function for the Pico TX alarm.
- *
- * This function is called when the Pico TX alarm is triggered.
- *
- * @param id The ID of the alarm.
- * @param user_data A pointer to user-defined data.
- * @return The return value is not used.
- */
-int64_t pico_tx_alarm_callback(alarm_id_t id, void *user_data);
-/**
- * @brief Callback function for the Pico RX repeating timer.
- *
- * This function is called when the Pico RX repeating timer expires.
- *
- * @param t Pointer to the repeating timer structure.
- * @return True if the callback was successful, false otherwise.
- */
-bool pico_rx_repeating_timer_callback(struct repeating_timer *t);
-
-/**
- * @brief Callback function for a repeating timer in the pico_tx module.
- *
- * This function is called when the repeating timer expires. It is responsible for handling the timer event
- * and performing any necessary actions.
- *
- * @param t A pointer to the repeating timer structure.
- * @return true if the timer event was handled successfully, false otherwise.
- */
-bool pico_tx_repeating_timer_callback(struct repeating_timer *t);
+void pico_init_transmitter(rf_pico_transmitter* self);
 
 /**
  * Sends a message using the RF Pico transmitter.
@@ -108,14 +56,7 @@ void pico_tx_send_message(rf_pico_transmitter* transmitter, RF_Message message);
  * @param user_data A pointer to user-defined data.
  * @return The return value is not used.
  */
-void pico_rx_start_receiving(rf_pico_receiver* self);
-
-/**
- * @brief Stops the Pico RX from receiving.
- *
- * @param self The RF Pico receiver.
- */
-void pico_rx_stop_receiving(rf_pico_receiver* self);
+void pico_init_receiver(rf_pico_receiver* self, void* result_callback);
 
 /**
  * @brief Callback function for the Pico RX alarm.
@@ -126,61 +67,13 @@ void pico_rx_stop_receiving(rf_pico_receiver* self);
  * @param user_data A pointer to user-defined data.
  * @return The return value is not used.
  */
-void pico_init_receiver(rf_pico_receiver* self, void* result_callback);
+void pico_rx_start_receiving(rf_pico_receiver* self);
 
 /**
- * @brief Initializes the RF Pico transmitter.
- *
- * @param self The RF Pico transmitter.
- * @param gpio The GPIO pin to use for the transmitter.
- */
-void pico_init_transmitter(rf_pico_transmitter* self);
-
-/**
- * @brief Initializes the RF Pico receiver.
+ * @brief Stops the Pico RX from receiving.
  *
  * @param self The RF Pico receiver.
- * @param gpio The GPIO pin to use for the receiver.
  */
-void pico_tx_set_onetime_trigger_time(uint64_t time_to_trigger, void* user_data);
-
-
-/**
- * Sets the recurring trigger time for pico_tx.
- *
- * This function sets the time at which the pico_tx will trigger periodically.
- *
- * @param time_to_trigger The time at which the trigger should occur.
- * @param user_data      Optional user data to be passed to the trigger callback.
- */
-void pico_tx_set_recurring_trigger_time(uint64_t time_to_trigger, void* user_data);
-
-/**
- * Cancels the recurring trigger for pico_tx.
- *
- * This function cancels the recurring trigger for pico_tx.
- *
- * @param user_data Optional user data to be passed to the trigger callback.
- */
-void pico_tx_cancel_trigger(void* user_data);
-
-/**
- * Sets the recurring trigger time for pico_rx.
- *
- * This function sets the time at which the pico_rx will trigger periodically.
- *
- * @param time_to_trigger The time at which the trigger should occur.
- * @param user_data      Optional user data to be passed to the trigger callback.
- */
-void pico_rx_set_recurring_trigger_time(uint64_t time_to_trigger, void* user_data);
-
-/**
- * Cancels the recurring trigger for pico_rx.
- *
- * This function cancels the recurring trigger for pico_rx.
- *
- * @param user_data Optional user data to be passed to the trigger callback.
- */
-void pico_rx_cancel_trigger(void* user_data);
+void pico_rx_stop_receiving(rf_pico_receiver* self);
 
 #endif // RF_PICO_H
